+++ src/components/grid-pagination/grid-pagination.tsx
@@ -18,9 +18,10 @@ import { GridPaginationArrow } from './grid-pagination-arrow';
     shadow: true
 })
 export class GridPagination implements ComponentInterface {
+    static readonly INDEX_FIRST_PAGE: number = 0;
     @Element() el!: HTMLWcsGridPaginationElement;
     @Prop() availablePageSizes: number[] = [10, 20, 50];
-    @Prop() currentPage: number = 0;
+    @Prop() currentPage: number = GridPagination.INDEX_FIRST_PAGE;
     @Prop() pageSize: number = this.availablePageSizes[0];
     @Prop() itemsCount: number = 0;
     @Prop() pageCount: number = 1;