+++ CHANGELOG.md
@@ -25,7 +25,7 @@ for now removed features.
 
 ### Fixed
 
-for any bug fixes.
+- **grid**: (front side pagination) when we reload less data than current data, we have to reset current page if it is greater than total of pages
 
 ### Security